Optimized codebase for aprgmath module

+++ AprgMath/AprgMath/src/Math/GameTheory/SpecificGames/GameWithMaze.cpp
@@ -84,17 +84,15 @@ GameWithMaze::Coordinate GameWithMaze::getNextCoordinateWithGrundyNumber(
 string GameWithMaze::getString() {
     DisplayTable table;
     table.setBorders("-", "|");
-    for (unsigned int y = 0; y < m_isBlockedMatrix.getNumberOfRows(); y++) {
+    for (int y = 0; y < static_cast<int>(m_isBlockedMatrix.getNumberOfRows()); y++) {
         table.addRow();
-        for (unsigned int x = 0; x < m_isBlockedMatrix.getNumberOfColumns(); x++) {
+        for (int x = 0; x < static_cast<int>(m_isBlockedMatrix.getNumberOfColumns()); x++) {
             stringstream ss;
             if (m_isBlockedMatrix.getEntry(x, y)) {
-                ss << "X";
-            } else {
+                ss << "X";            } else {
                 ss << getGrundyNumberAt(Coordinate(x, y));
             }
-            table.getLastRow().addCell(ss.str());
-        }
+            table.getLastRow().addCell(ss.str());        }
     }
     stringstream ss;
     ss << "Matrix output:\n" << table;
@@ -120,27 +118,24 @@ GameWithMaze::Coordinates GameWithMaze::getNextCoordinates(Coordinate const& coo
 }
 
 void GameWithMaze::retrieveLeftCoordinates(Coordinates& retrievedCoordinates, Coordinate const& coordinate) const {
-    for (int x = static_cast<int>(coordinate.first) - 1; x >= 0; x--) {
-        Coordinate xyToCheck(static_cast<unsigned int>(x), coordinate.second);
+    for (int x = coordinate.first - 1; x >= 0; x--) {
+        Coordinate xyToCheck(x, coordinate.second);
         if (!m_isBlockedMatrix.getEntry(xyToCheck.first, xyToCheck.second)) {
             retrievedCoordinates.emplace_back(xyToCheck);
-        } else {
-            break;
+        } else {            break;
         }
     }
 }
 
 void GameWithMaze::retrieveUpCoordinates(Coordinates& retrievedCoordinates, Coordinate const& coordinate) const {
-    for (int y = static_cast<int>(coordinate.second) - 1; y >= 0; y--) {
-        Coordinate xyToCheck(coordinate.first, static_cast<unsigned int>(y));
+    for (int y = coordinate.second - 1; y >= 0; y--) {
+        Coordinate xyToCheck(coordinate.first, y);
         if (!m_isBlockedMatrix.getEntry(xyToCheck.first, xyToCheck.second)) {
             retrievedCoordinates.emplace_back(xyToCheck);
-        } else {
-            break;
+        } else {            break;
         }
     }
 }
-
 }  // namespace math
 
 }  // namespace alba
