+++ AprgMath/AprgMath/src/Math/NumberTheory/ModularArithmetic.cpp
@@ -55,15 +55,13 @@ bool canModuloBeDistributedInEveryIterationOfFactorial(UnsignedInteger const num
 
     UnsignedInteger undistributed = getFactorial(number) % modulo;
     UnsignedInteger distributed = 1;
-    for (unsigned int i = 2; i <= number; i++) {
+    for (int i = 2; i <= static_cast<int>(number); i++) {
         distributed = (distributed * i) % modulo;
     }
-    distributed %= modulo;
-    return undistributed == distributed;
+    distributed %= modulo;    return undistributed == distributed;
 }
 
-bool isFermatTheoremTrue(UnsignedInteger const prime, UnsignedInteger const coPrime) {
-    // Fermat's theorem states that: (x^(m-1)) mod m = 1
+bool isFermatTheoremTrue(UnsignedInteger const prime, UnsignedInteger const coPrime) {    // Fermat's theorem states that: (x^(m-1)) mod m = 1
     // where m is prime and x and m are coprime.
 
     bool result(false);
@@ -195,13 +193,11 @@ UnsignedInteger getModularFactorial(UnsignedInteger const number, UnsignedIntege
     // n! mod m = ((n mod m) * (n-1 mod m) * (n-2 mod m) ... (1 mod m)) mod m
 
     UnsignedInteger result = 1;
-    for (unsigned int i = 2; i <= number; i++) {
+    for (int i = 2; i <= static_cast<int>(number); i++) {
         result = (result * i) % modulo;
     }
-    result %= modulo;
-    return result;
+    result %= modulo;    return result;
 }
 
 }  // namespace math
-
 }  // namespace alba
